@@ -1,59 +1,60 @@
 using System;
-using System.Linq;
+using System.Globalization;
 
-namespace SmartDateRangeParser
+namespace SmartDateRangeParser;
+
+public readonly struct SmartDateRange(DateTime start, DateTime end)
 {
-    public class SmartDateRange
+    public DateTime Start { get; } = start;
+    public DateTime End { get; } = end;
+
+    public static SmartDateRange Parse(string input)
     {
-        public DateTime Start { get; }
-        public DateTime End { get; }
+        ArgumentException.ThrowIfNullOrWhiteSpace(input);
 
-        private SmartDateRange(DateTime start, DateTime end)
+        input = input.Trim().ToLowerInvariant();
+
+        return input switch
         {
-            Start = start;
-            End = end;
+            "today" => new(DateTime.Today, DateTime.Today),
+            "yesterday" => new(DateTime.Today.AddDays(-1), DateTime.Today.AddDays(-1)),
+            _ when input.StartsWith("last ") && input.Contains("business day", StringComparison.OrdinalIgnoreCase)
+                => ParseLastBusinessDays(input),
+            _ => throw new NotSupportedException($"Unsupported date range expression: '{input}'")
+        };
+    }
+
+    private static SmartDateRange ParseLastBusinessDays(string input)
+    {
+        int days = ExtractNumber(input);
+        if (days <= 0)
+            throw new ArgumentOutOfRangeException(nameof(days), "Number of business days must be greater than 0.");
+
+        var end = DateTime.Today.AddDays(-1);
+        var start = SubtractBusinessDays(end, days - 1);
+        return new(start, end);
+    }
+
+    private static int ExtractNumber(string input)
+    {
+        foreach (var token in input.Split(' '))
+        {
+            if (int.TryParse(token, NumberStyles.Integer, CultureInfo.InvariantCulture, out int number))
+                return number;
         }
 
-        public static SmartDateRange Parse(string input)
+        throw new FormatException("Could not find a valid number in input.");
+    }
+
+    private static DateTime SubtractBusinessDays(DateTime from, int count)
+    {
+        while (count > 0)
         {
-            input = input.ToLowerInvariant();
-
-            if (input == "today")
-            {
-                var today = DateTime.Today;
-                return new SmartDateRange(today, today);
-            }
-
-            if (input.StartsWith("last "))
-            {
-                if (input.Contains("business days"))
-                {
-                    int days = ExtractNumber(input);
-                    var end = DateTime.Today.AddDays(-1);
-                    var start = SubtractBusinessDays(end, days - 1);
-                    return new SmartDateRange(start, end);
-                }
-            }
-
-            throw new NotSupportedException($"Unsupported format: {input}");
+            from = from.AddDays(-1);
+            if (from.DayOfWeek is not DayOfWeek.Saturday and not DayOfWeek.Sunday)
+                count--;
         }
 
-        private static int ExtractNumber(string input)
-        {
-            return int.TryParse(new string(input.Where(char.IsDigit).ToArray()), out int result)
-                ? result
-                : throw new FormatException("Could not extract number from input.");
-        }
-
-        private static DateTime SubtractBusinessDays(DateTime from, int days)
-        {
-            while (days > 0)
-            {
-                from = from.AddDays(-1);
-                if (from.DayOfWeek != DayOfWeek.Saturday && from.DayOfWeek != DayOfWeek.Sunday)
-                    days--;
-            }
-            return from;
-        }
+        return from;
     }
 }
